requirements.txt
Flask
tensorflow
tensorflow-datasets
google-generativeai
python-dotenv
pillow
werkzeug
